Use `modals` for verification
