+++ package.json
@@ -21,7 +21,7 @@
     "@discordjs/rest": "^0.4.1",
     "colorts": "^0.1.63",
     "discord-api-types": "^0.31.2",
-    "discord.js": "^13.6.0",
+    "discord.js": "^13.7.0",
     "node-tesseract-ocr": "^2.2.1",
     "ts-node-dev": "^1.1.8",
     "typescript": "^4.6.3"
